+++ .vscode/settings.json
@@ -1,3 +1,2 @@
 {
-    "python.pythonPath": "env32\\Scripts\\python.exe"
 }
