+++ server_monitor.py
@@ -1,0 +1,21 @@
+import sys
+import subprocess
+import socket
+import os
+
+
+'''
+class system_info:
+
+  def 
+
+'''
+
+
+'''
+class monitor:
+
+  def
+
+
+'''
